feat(core): add checkHealth extention to model core
Extending from the healthcheck for SQLite core, the models init()
method is where a checkHealth() can be called during the bootstrap
of the core runtime. That is, before allowing the server to run
with misconfiguration, the healthcheck will provide assurance that
the server is correctly initialized with whichever Model provider
is, as long as the function is implemented properly.

Note that the init() method is defined, but never called during the
runtime bootstrap process.